requirements.txt
aiogram==3.7.0
asyncpg==0.29.0
APScheduler==3.10.4
python-dotenv==1.0.1
openpyxl==3.1.2
pytz==2024.1
types-pytz==2024.1.0.20240203
